export ValidDateFromSelf and rename validDate filter
