+++ .changeset/thick-cooks-tap.md
@@ -1,0 +1,5 @@
+---
+"@effect/schema": patch
+---
+
+export ValidDateFromSelf and rename validDate filter
